+++ ta4j-core/src/test/java/org/ta4j/core/analysis/criteria/NumberOfBarsCriterionTest.java
@@ -24,13 +24,18 @@
 package org.ta4j.core.analysis.criteria;
 
 import org.junit.Test;
-import org.ta4j.core.*;
+import org.ta4j.core.AnalysisCriterion;
+import org.ta4j.core.BaseTradingRecord;
+import org.ta4j.core.Order;
+import org.ta4j.core.Trade;
+import org.ta4j.core.TradingRecord;
 import org.ta4j.core.mocks.MockTimeSeries;
 import org.ta4j.core.num.Num;
 
 import java.util.function.Function;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 import static org.ta4j.core.TestUtils.assertNumEquals;
 
 public class NumberOfBarsCriterionTest extends AbstractCriterionTest {
@@ -51,8 +56,8 @@ public class NumberOfBarsCriterionTest extends AbstractCriterionTest {
     public void calculateWithTwoTrades() {
         MockTimeSeries series = new MockTimeSeries(numFunction, 100, 105, 110, 100, 95, 105);
         TradingRecord tradingRecord = new BaseTradingRecord(
-                Order.buyAt(0,series), Order.sellAt(2,series),
-                Order.buyAt(3,series), Order.sellAt(5,series));
+                Order.buyAt(0, series), Order.sellAt(2, series),
+                Order.buyAt(3, series), Order.sellAt(5, series));
 
         AnalysisCriterion numberOfBars = getCriterion();
         assertNumEquals(6, numberOfBars.calculate(series, tradingRecord));
@@ -61,7 +66,7 @@ public class NumberOfBarsCriterionTest extends AbstractCriterionTest {
     @Test
     public void calculateWithOneTrade() {
         MockTimeSeries series = new MockTimeSeries(numFunction, 100, 95, 100, 80, 85, 70);
-        Trade t = new Trade(Order.buyAt(2,series), Order.sellAt(5,series));
+        Trade t = new Trade(Order.buyAt(2, series), Order.sellAt(5, series));
         AnalysisCriterion numberOfBars = getCriterion();
         assertNumEquals(4, numberOfBars.calculate(series, t));
     }
@@ -71,5 +76,10 @@ public class NumberOfBarsCriterionTest extends AbstractCriterionTest {
         AnalysisCriterion criterion = getCriterion();
         assertTrue(criterion.betterThan(numOf(3), numOf(6)));
         assertFalse(criterion.betterThan(numOf(6), numOf(2)));
+    }
+
+    @Test
+    public void testCalculateOneOpenTradeShouldReturnZero() {
+        openedTradeUtils.testCalculateOneOpenTradeShouldReturnExpectedValue(numFunction, getCriterion(), 0);
     }
 }