+++ ta4j-core/src/test/java/org/ta4j/core/analysis/criteria/TotalProfitCriterionTest.java
@@ -24,16 +24,21 @@
 package org.ta4j.core.analysis.criteria;
 
 import org.junit.Test;
-import org.ta4j.core.*;
+import org.ta4j.core.AnalysisCriterion;
+import org.ta4j.core.BaseTradingRecord;
+import org.ta4j.core.Order;
+import org.ta4j.core.Trade;
+import org.ta4j.core.TradingRecord;
 import org.ta4j.core.mocks.MockTimeSeries;
 import org.ta4j.core.num.Num;
 
 import java.util.function.Function;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 import static org.ta4j.core.TestUtils.assertNumEquals;
 
-public class TotalProfitCriterionTest extends AbstractCriterionTest{
+public class TotalProfitCriterionTest extends AbstractCriterionTest {
 
     public TotalProfitCriterionTest(Function<Number, Num> numFunction) {
         super((params) -> new TotalProfitCriterion(), numFunction);
@@ -43,8 +48,8 @@ public class TotalProfitCriterionTest extends AbstractCriterionTest{
     public void calculateOnlyWithGainTrades() {
         MockTimeSeries series = new MockTimeSeries(numFunction, 100, 105, 110, 100, 95, 105);
         TradingRecord tradingRecord = new BaseTradingRecord(
-                Order.buyAt(0,series), Order.sellAt(2,series),
-                Order.buyAt(3,series), Order.sellAt(5,series));
+                Order.buyAt(0, series), Order.sellAt(2, series),
+                Order.buyAt(3, series), Order.sellAt(5, series));
 
         AnalysisCriterion profit = getCriterion();
         assertNumEquals(1.10 * 1.05, profit.calculate(series, tradingRecord));
@@ -54,8 +59,8 @@ public class TotalProfitCriterionTest extends AbstractCriterionTest{
     public void calculateOnlyWithLossTrades() {
         MockTimeSeries series = new MockTimeSeries(numFunction, 100, 95, 100, 80, 85, 70);
         TradingRecord tradingRecord = new BaseTradingRecord(
-                Order.buyAt(0,series), Order.sellAt(1,series),
-                Order.buyAt(2,series), Order.sellAt(5,series));
+                Order.buyAt(0, series), Order.sellAt(1, series),
+                Order.buyAt(2, series), Order.sellAt(5, series));
 
         AnalysisCriterion profit = getCriterion();
         assertNumEquals(0.95 * 0.7, profit.calculate(series, tradingRecord));
@@ -65,8 +70,8 @@ public class TotalProfitCriterionTest extends AbstractCriterionTest{
     public void calculateProfitWithTradesThatStartSelling() {
         MockTimeSeries series = new MockTimeSeries(numFunction, 100, 95, 100, 80, 85, 70);
         TradingRecord tradingRecord = new BaseTradingRecord(
-                Order.sellAt(0,series), Order.buyAt(1,series),
-                Order.sellAt(2,series), Order.buyAt(5,series));
+                Order.sellAt(0, series), Order.buyAt(1, series),
+                Order.sellAt(2, series), Order.buyAt(5, series));
 
         AnalysisCriterion profit = getCriterion();
         assertNumEquals((1 / 0.95) * (1 / 0.7), profit.calculate(series, tradingRecord));
@@ -95,5 +100,10 @@ public class TotalProfitCriterionTest extends AbstractCriterionTest{
         AnalysisCriterion criterion = getCriterion();
         assertTrue(criterion.betterThan(numOf(2.0), numOf(1.5)));
         assertFalse(criterion.betterThan(numOf(1.5), numOf(2.0)));
+    }
+
+    @Test
+    public void testCalculateOneOpenTradeShouldReturnOne() {
+        openedTradeUtils.testCalculateOneOpenTradeShouldReturnExpectedValue(numFunction, getCriterion(), 1);
     }
 }